// shared/contracts/generated/Dsp.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { DspStatus } from "./DspStatus";
import type { Environment } from "./Environment";

export type Dsp = { id: string, name: string, environment: Environment, status: DspStatus, timezone: string, permanent: boolean, revision: number, createdAt: string, };